+++ CHANGELOG.md
@@ -2,5 +2,6 @@ hateoas-browser CHANGELOG
 =========================
 
 * dev
+  - added headers for request
   - added panels and tabs to split result
   - initial version